pillboxtracker/requirements.txt
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==3.2.16
django-appconf==1.1.0
django-bootstrap5==22.2
django_debug_toolbar==3.8.1
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kombu==5.5.4
packaging==25.0
Pillow==9.3.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==22.3
pytz==2025.2
redis==6.3.0
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
